# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marginalia - Social Reading Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type BookStatus = (typeof BookStatus)[keyof typeof BookStatus];

export const BookStatus = {
  reading: "reading",
  completed: "completed",
  wishlist: "wishlist",
} as const;

export interface Book {
  id: number;
  title: string;
  author: string;
  progress: number;
  currentPage: number;
  totalPages: number;
  annotations: number;
  highlights: number;
  debates: number;
  currentChapter: string;
  status: BookStatus;
}

export interface HeatmapEntry {
  chapter: number;
  chapterTitle: string;
  intensity: number;
  annotationCount: number;
  locked: boolean;
}

export type BookDetailStatus =
  (typeof BookDetailStatus)[keyof typeof BookDetailStatus];

export const BookDetailStatus = {
  reading: "reading",
  completed: "completed",
  wishlist: "wishlist",
} as const;

export type AnnotationType =
  (typeof AnnotationType)[keyof typeof AnnotationType];

export const AnnotationType = {
  insight: "insight",
  reaction: "reaction",
  theory: "theory",
  highlight: "highlight",
} as const;

export type AnnotationReactions = { [key: string]: number };

export interface Annotation {
  id: number;
  bookId: number;
  bookTitle: string;
  chapter: string;
  progressAt: number;
  excerpt: string;
  /** @nullable */
  note?: string | null;
  type: AnnotationType;
  userId: string;
  userName: string;
  userInitials: string;
  createdAt: string;
  reactions: AnnotationReactions;
  replyCount: number;
  isPublic: boolean;
}

export interface BookDetail {
  id: number;
  title: string;
  author: string;
  progress: number;
  currentPage: number;
  totalPages: number;
  annotations: number;
  highlights: number;
  debates: number;
  currentChapter: string;
  status: BookDetailStatus;
  heatmap: HeatmapEntry[];
  recentAnnotations: Annotation[];
}

export type ThreadReplyReactions = { [key: string]: number };

export interface ThreadReply {
  id: number;
  userId: string;
  userName: string;
  userInitials: string;
  text: string;
  createdAt: string;
  reactions: ThreadReplyReactions;
}

export interface AnnotationThread {
  annotation: Annotation;
  replies: ThreadReply[];
}

export type CreateAnnotationBodyType =
  (typeof CreateAnnotationBodyType)[keyof typeof CreateAnnotationBodyType];

export const CreateAnnotationBodyType = {
  insight: "insight",
  reaction: "reaction",
  theory: "theory",
  highlight: "highlight",
} as const;

export interface CreateAnnotationBody {
  bookId: number;
  chapter: string;
  progressAt: number;
  excerpt: string;
  /** @nullable */
  note?: string | null;
  type: CreateAnnotationBodyType;
  isPublic: boolean;
}

export interface AddReactionBody {
  reaction: string;
}

export interface FeedResponse {
  currentReading?: BookDetail;
  echoes: Annotation[];
  discover: Book[];
}

export interface UserProfile {
  id: string;
  name: string;
  initials: string;
  /** @nullable */
  bio?: string | null;
  booksRead: number;
  totalAnnotations: number;
  totalHighlights: number;
  identityPhrase: string;
  recentExcerpts: Annotation[];
}

export interface ExploreResponse {
  trending: Annotation[];
  readers: UserProfile[];
  emergingBooks: Book[];
}

export type ShareableCardType =
  (typeof ShareableCardType)[keyof typeof ShareableCardType];

export const ShareableCardType = {
  moment: "moment",
  insight: "insight",
  reaction: "reaction",
  progress: "progress",
  identity: "identity",
} as const;

export type ShareableCardStats = { [key: string]: number };

export interface ShareableCard {
  id: number;
  bookId: number;
  type: ShareableCardType;
  title: string;
  content: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  chapter?: string | null;
  /** @nullable */
  progressAt?: number | null;
  stats?: ShareableCardStats;
}