# lib/api-zod/src/generated/types/annotation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marginalia - Social Reading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnotationReactions } from "./annotationReactions";
import type { AnnotationType } from "./annotationType";

export interface Annotation {
  id: number;
  bookId: number;
  bookTitle: string;
  chapter: string;
  progressAt: number;
  excerpt: string;
  /** @nullable */
  note?: string | null;
  type: AnnotationType;
  userId: string;
  userName: string;
  userInitials: string;
  createdAt: string;
  reactions: AnnotationReactions;
  replyCount: number;
  isPublic: boolean;
}