# lib/api-zod/src/generated/types/bookDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marginalia - Social Reading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Annotation } from "./annotation";
import type { BookDetailStatus } from "./bookDetailStatus";
import type { HeatmapEntry } from "./heatmapEntry";

export interface BookDetail {
  id: number;
  title: string;
  author: string;
  progress: number;
  currentPage: number;
  totalPages: number;
  annotations: number;
  highlights: number;
  debates: number;
  currentChapter: string;
  status: BookDetailStatus;
  heatmap: HeatmapEntry[];
  recentAnnotations: Annotation[];
}