# lib/api-zod/src/generated/types/createAnnotationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marginalia - Social Reading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAnnotationBodyType } from "./createAnnotationBodyType";

export interface CreateAnnotationBody {
  bookId: number;
  chapter: string;
  progressAt: number;
  excerpt: string;
  /** @nullable */
  note?: string | null;
  type: CreateAnnotationBodyType;
  isPublic: boolean;
}