# lib/api-zod/src/generated/types/exploreResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marginalia - Social Reading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Annotation } from "./annotation";
import type { Book } from "./book";
import type { UserProfile } from "./userProfile";

export interface ExploreResponse {
  trending: Annotation[];
  readers: UserProfile[];
  emergingBooks: Book[];
}