# lib/api-zod/src/generated/types/heatmapEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marginalia - Social Reading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface HeatmapEntry {
  chapter: number;
  chapterTitle: string;
  intensity: number;
  annotationCount: number;
  locked: boolean;
}