# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marginalia - Social Reading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Annotation } from "./annotation";

export interface UserProfile {
  id: string;
  name: string;
  initials: string;
  /** @nullable */
  bio?: string | null;
  booksRead: number;
  totalAnnotations: number;
  totalHighlights: number;
  identityPhrase: string;
  recentExcerpts: Annotation[];
}